crimson/requirement.txt
certifi==2019.9.11
chardet==3.0.4
Django==2.2.7
django-file-download==1.0.2
django-json-response==1.1.5
django-logit==1.1.3
django-paypal==1.0.0
django-response==1.1.1
idna==2.8
pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
TimeConvert==1.5.0
tzlocal==2.0.0
urllib3==1.25.7
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.3.3
Automat==0.7.0
autopep8==1.5
cachetools==3.1.0
certifi==2019.9.11
channels==2.2.0
chardet==3.0.4
constantly==15.1.0
daphne==2.3.0
defusedxml==0.6.0
Django==2.2.1
django-allauth==0.41.0
django-crispy-forms==1.7.2
django-file-download==1.0.2
django-filter==2.1.0
django-json-response==1.1.5
django-logit==1.1.3
django-paypal==1.0.0
django-recaptcha2==1.4.1
django-response==1.1.1
djangorestframework==3.9.4
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
httplib2==0.12.3
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
Markdown==3.1
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==6.0.0
pkg-resources==0.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
PyHamcrest==1.9.0
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
TimeConvert==1.5.0
Twisted==19.2.0
txaio==18.8.1
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.7
zope.interface==4.6.0
